+++ internal/config/config.go
@@ -42,15 +42,17 @@ type ReverseCreateCheckOptions struct {
 
 // Check defines the check parameters
 type Check struct {
-	Broker           string `json:"broker" yaml:"broker" toml:"broker"`
-	BundleID         string `mapstructure:"bundle_id" json:"bundle_id" yaml:"bundle_id" toml:"bundle_id"`
-	Create           bool   `mapstructure:"create" json:"create" yaml:"create" toml:"create"`
-	MetricFilters    string `mapstructure:"metric_filters" json:"metric_filters" yaml:"metric_filters" toml:"metric_filters"` // needs to be json embedded in a string because rules are positional
-	MetricRefreshTTL string `mapstructure:"metric_refresh_ttl" json:"metric_refresh_ttl" yaml:"metric_refresh_ttl" toml:"metric_refresh_ttl"`
-	MetricStreamtags bool   `mapstructure:"metric_streamtags" json:"metric_streamtags" yaml:"metric_streamtags" toml:"metric_streamtags"`
-	Tags             string `json:"tags" yaml:"tags" toml:"tags"`
-	Target           string `mapstructure:"target" json:"target" yaml:"target" toml:"target"`
-	Title            string `json:"title" yaml:"title" toml:"title"`
+	Broker              string `json:"broker" yaml:"broker" toml:"broker"`
+	BundleID            string `mapstructure:"bundle_id" json:"bundle_id" yaml:"bundle_id" toml:"bundle_id"`
+	Create              bool   `mapstructure:"create" json:"create" yaml:"create" toml:"create"`
+	MetricFilters       string `mapstructure:"metric_filters" json:"metric_filters" yaml:"metric_filters" toml:"metric_filters"` // needs to be json embedded in a string because rules are positional
+	MetricFilterFile    string `mapstructure:"metric_filter_file" json:"metric_filter_file" yaml:"metric_filter_file" toml:"metric_filter_file"`
+	UpdateMetricFilters bool   `mapstructure:"update_metric_filters" json:"update_metric_filters" yaml:"update_metric_filters" toml:"update_metric_filters"`
+	MetricRefreshTTL    string `mapstructure:"metric_refresh_ttl" json:"metric_refresh_ttl" yaml:"metric_refresh_ttl" toml:"metric_refresh_ttl"`
+	MetricStreamtags    bool   `mapstructure:"metric_streamtags" json:"metric_streamtags" yaml:"metric_streamtags" toml:"metric_streamtags"`
+	Tags                string `json:"tags" yaml:"tags" toml:"tags"`
+	Target              string `mapstructure:"target" json:"target" yaml:"target" toml:"target"`
+	Title               string `json:"title" yaml:"title" toml:"title"`
 	// hide deprecated config settings
 	EnableNewMetrics bool   `json:"-" yaml:"-" toml:"-"`
 	MetricStateDir   string `json:"-" yaml:"-" toml:"-"`
@@ -282,6 +284,10 @@ const (
 	// TOML configuration file:
 	//  `metric_filters = '''[["deny","$^",""],["allow","^.+$",""]]'''`
 	KeyCheckMetricFilters = "check.metric_filters"
+	// KeyCheckMetricFilterFile an external JSON file to use (see etc/example_metric_filters.json)
+	KeyCheckMetricFilterFile = "check.metric_filter_file"
+	// KeyCheckUpdateMetricFIlters force update of check with configured metric filters on start
+	KeyCheckUpdateMetricFilters = "check.update_metric_filters"
 
 	// KeyCheckCreate toggles creating a new check bundle when a check bundle id is not supplied
 	KeyCheckCreate = "check.create"